add CHANNEL_UPDATE
